+++ cli/cli/Utils/LogConfigData.cs
@@ -1,7 +1,0 @@
-namespace cli.Utils;
-
-public class LogConfigData
-{
-	public bool shouldUseTempFile;
-	public string logFilePath;
-}