+++ cli/cli/CHANGELOG.md
@@ -5,10 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.1.0]
+
+### Added
+- `--unmask-logs` option will show full tokens in verbose logs
+- `--no-log-file` option will prevent verbose logs from being written to temp file
 
 ### Added
 - `beam project enable` and `beam project disable` commands will set the `<BeamEnabled>` setting.
+- Can pass MSBuild dlls location through environment variable to the CLI
+
 
 ## [2.0.1] - 2024-06-17